Add ability to skip certain test variants

+++ scripts/avd_test.sh
@@ -171,17 +171,21 @@ test_main() {
     emu_args="$emu_args -show-kernel -logcat '' -logcat-output logcat.log"
   fi
 
-  # Patch and test debug build
-  ./build.py avd_patch -s "$ramdisk" magisk_patched.img
-  kill -INT $emu_pid
-  wait $emu_pid
-  test_emu debug $api
+  if [ -z "$AVD_TEST_SKIP_DEBUG" ]; then
+    # Patch and test debug build
+    ./build.py avd_patch -s "$ramdisk" magisk_patched.img
+    kill -INT $emu_pid
+    wait $emu_pid
+    test_emu debug $api
+  fi
 
-  # Patch and test release build
-  ./build.py -r avd_patch -s "$ramdisk" magisk_patched.img
-  kill -INT $emu_pid
-  wait $emu_pid
-  test_emu release $api
+  if [ -z "$AVD_TEST_SKIP_RELEASE" ]; then
+    # Patch and test release build
+    ./build.py -r avd_patch -s "$ramdisk" magisk_patched.img
+    kill -INT $emu_pid
+    wait $emu_pid
+    test_emu release $api
+  fi
 
   # Cleanup
   kill -INT $emu_pid
